feat(ScrollView): promote fragment to component (#8981)
ScrollView is a general-purpose scrolling primitive already used across
Drawer, Dialog, Table, List, Filter, and Forms. Promoting it to a
component makes the public API easier to discover and supports reuse in
other projects.

The existing `@dnb/eufemia/fragments` import remains as a deprecated
compatibility export. It is scheduled for removal in v13, where
consumers should use `@dnb/eufemia` instead. Specialized
`Table.ScrollView` and `List.ScrollView` variants remain available for
their component-specific behavior.
