+++ packages/dnb-design-system-portal/src/docs/EUFEMIA_CHANGELOG.mdx
@@ -1,3 +1,7 @@
+## August, 14. 2026
+
+- New component: [ScrollView](/uilib/components/scroll-view) for horizontal and vertical scrolling areas. The previous import from `@dnb/eufemia/fragments` is deprecated and will be removed in v13.
+
 ## June, 11. 2026
 
 - New component: [Filter](/uilib/components/filter) for composable, data-model-agnostic filter UIs with shared state via hooks.
