+++ packages/dnb-design-system-portal/src/docs/uilib/components/fragments.mdx
@@ -15,11 +15,7 @@ import ListFragments from 'dnb-design-system-portal/src/shared/parts/ListFragmen
 You import them like so:
 
 ```jsx
-import {
-  DrawerList,
-  ScrollView,
-  TextCounter,
-} from '@dnb/eufemia/fragments'
+import { DrawerList, TextCounter } from '@dnb/eufemia/fragments'
 ```
 
 ## Description
@@ -27,6 +23,8 @@ import {
 Fragments are small, low-level and reusable parts used inside other components.
 
 You may use them only to build new components from.
+
+The deprecated `ScrollView` fragment export remains available for backwards compatibility until v13. Import the [ScrollView component](/uilib/components/scroll-view/) from `@dnb/eufemia` instead.
 
 ## Available Fragments
