--- packages/dnb-design-system-portal/src/docs/uilib/components/fragments/scroll-view/demos.mdx
@@ -1,15 +1,0 @@
----
-showTabs: true
----
-
-import * as Examples from './Examples'
-
-## Demos
-
-### Keyboard support
-
-When used for regular content, it should be possible for the user to user their keyboard to control the scroll position.
-
-You can enable keyboard support with the `interactive` property.
-
-<Examples.ScrollViewInteractive />
