--- packages/dnb-design-system-portal/src/docs/uilib/components/fragments/scroll-view/info.mdx
@@ -1,34 +1,0 @@
----
-showTabs: true
----
-
-import { ScrollViewInfo } from 'Docs/uilib/components/fragments/scroll-view/Examples'
-
-## Import
-
-```tsx
-import { ScrollView } from '@dnb/eufemia/fragments'
-```
-
-## Description
-
-ScrollView is a tiny helper component helping the user control overflowing content horizontally or vertically.
-
-It also is used in other floating components like [Dropdown](/uilib/components/dropdown) or [Drawer](/uilib/components/drawer).
-
-## Relevant links
-
-- [Source code](https://github.com/dnbexperience/eufemia/tree/main/packages/dnb-eufemia/src/components/fragments/scroll-view)
-- [Docs code](https://github.com/dnbexperience/eufemia/tree/main/packages/dnb-design-system-portal/src/docs/uilib/components/fragments/scroll-view)
-
-### Scrollbar gutter
-
-Use `scrollbarGutter="stable"` to reserve space for the scrollbar, preventing horizontal layout shifts when content dynamically changes between overflowing and non-overflowing states.
-
-This uses the CSS `scrollbar-gutter: stable` property. On systems with overlay scrollbars (such as iOS or macOS default), it has no visible effect. On systems with classic scrollbars (such as Windows and Linux), it prevents content from shifting when a scrollbar appears or disappears.
-
-[Dialog](/uilib/components/dialog) and [Drawer](/uilib/components/drawer) enable this automatically by default.
-
-<ScrollViewInfo />
-
-<RelatedComponents />
