+++ packages/dnb-design-system-portal/src/docs/uilib/components/scroll-view.mdx
@@ -4,12 +4,14 @@ category: 'other'
 description: 'Use ScrollView when content needs its own horizontal or vertical scrolling area.'
 showTabs: true
 theme: ['sbanken', 'carnegie']
+redirect_from:
+  - /uilib/components/fragments/scroll-view
 hideTabs:
   - title: Events
 ---
 
-import ScrollViewInfo from 'Docs/uilib/components/fragments/scroll-view/info'
-import ScrollViewDemos from 'Docs/uilib/components/fragments/scroll-view/demos'
+import ScrollViewInfo from 'Docs/uilib/components/scroll-view/info'
+import ScrollViewDemos from 'Docs/uilib/components/scroll-view/demos'
 
 <ScrollViewInfo />
 <ScrollViewDemos />
