+++ packages/dnb-design-system-portal/src/docs/uilib/components/scroll-view/demos.mdx
@@ -1,0 +1,13 @@
+---
+showTabs: true
+---
+
+import * as Examples from './Examples'
+
+## Demos
+
+### Keyboard-accessible overflow
+
+Use `interactive="auto"` to make the region keyboard-focusable only when its content overflows. The example constrains its height to create vertical overflow.
+
+<Examples.ScrollViewInteractive />
