+++ packages/dnb-design-system-portal/src/docs/uilib/components/scroll-view/properties.mdx
@@ -3,7 +3,7 @@ showTabs: true
 ---
 
 import PropertiesTable from 'dnb-design-system-portal/src/shared/parts/PropertiesTable'
-import { ScrollViewProperties } from '@dnb/eufemia/src/fragments/scroll-view/ScrollViewDocs'
+import { ScrollViewProperties } from '@dnb/eufemia/src/components/scroll-view/ScrollViewDocs'
 
 ## Properties
