+++ packages/dnb-eufemia/src/components/ScrollView.ts
@@ -1,0 +1,14 @@
+/**
+ * ATTENTION: This file is auto generated by using "prepareTemplates".
+ * Do not change the content!
+ *
+ */
+
+/**
+ * Library Index scroll-view to autogenerate all the components and extensions
+ * Used by "prepareScrollViews"
+ */
+
+import ScrollView from './scroll-view/ScrollView'
+export * from './scroll-view/ScrollView'
+export default ScrollView
