+++ packages/dnb-eufemia/src/components/drawer/DrawerContent.tsx
@@ -8,7 +8,7 @@ import type { JSX, ReactElement } from 'react'
 import { clsx } from 'clsx'
 import { findElementInChildren } from '../../shared/component-helper'
 import { getOffsetTop, warn } from '../../shared/helpers'
-import ScrollView from '../../fragments/scroll-view/ScrollView'
+import ScrollView from '../scroll-view/ScrollView'
 import DrawerHeader from './parts/DrawerHeader'
 import DrawerNavigation from './parts/DrawerNavigation'
 import ModalContext from '../modal/ModalContext'
