+++ packages/dnb-eufemia/src/components/list/ListScrollView.tsx
@@ -1,8 +1,8 @@
 import { useCallback, useContext, useRef, useState } from 'react'
 import type { CSSProperties, HTMLAttributes, ReactNode } from 'react'
 import { clsx } from 'clsx'
-import type { ScrollViewAllProps } from '../../fragments/scroll-view/ScrollView'
-import ScrollView from '../../fragments/scroll-view/ScrollView'
+import type { ScrollViewAllProps } from '../scroll-view/ScrollView'
+import ScrollView from '../scroll-view/ScrollView'
 import { useIsomorphicLayoutEffect as useLayoutEffect } from '../../shared/helpers/useIsomorphicLayoutEffect'
 import useCombinedRef from '../../shared/helpers/useCombinedRef'
 import { ListContext } from './ListContext'
