+++ packages/dnb-eufemia/src/components/scroll-view/__tests__/shared-mocks/ResizeObserver.ts
