+++ packages/dnb-eufemia/src/components/scroll-view/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Component Entry
+ *
+ */
+
+import ScrollView from './ScrollView'
+export default ScrollView
+export * from './ScrollView'
