+++ packages/dnb-eufemia/src/components/scroll-view/style.ts
@@ -1,0 +1,6 @@
+/**
+ * Web Style Import
+ *
+ */
+
+import './style/dnb-scroll-view.scss'
