+++ packages/dnb-eufemia/src/components/scroll-view/style/deps.scss
@@ -1,0 +1,6 @@
+/*
+* Used for snapshot testing
+*
+*/
+
+@use './dnb-scroll-view.scss';
