+++ packages/dnb-eufemia/src/components/scroll-view/style/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Web Style Import
+ *
+ */
+
+import './dnb-scroll-view.scss'
