+++ packages/dnb-eufemia/src/extensions/forms/Field/MultiSelection/MultiSelectionSelectedTags.tsx
@@ -1,7 +1,7 @@
 import { Fragment } from 'react'
 import { Accordion, Button, Flex, Tag } from '../../../../components'
 import { Hr, P } from '../../../../elements'
-import ScrollView from '../../../../fragments/scroll-view/ScrollView'
+import ScrollView from '../../../../components/scroll-view/ScrollView'
 import { close } from '../../../../icons'
 import type { MultiSelectionItem } from './MultiSelection'
