+++ packages/dnb-eufemia/src/fragments/scroll-view/index.ts
@@ -1,8 +1,5 @@
 /**
- * Component Entry
- *
+ * @deprecated Import `ScrollView` from `@dnb/eufemia` instead. This compatibility export will be removed in v13.
  */
-
-import ScrollView from './ScrollView'
-export default ScrollView
-export * from './ScrollView'
+export { default } from '../../components/scroll-view/ScrollView'
+export * from '../../components/scroll-view/ScrollView'
