+++ packages/dnb-eufemia/src/fragments/scroll-view/style.ts
@@ -1,6 +1,2 @@
-/**
- * Web Style Import
- *
- */
-
-import './style/dnb-scroll-view.scss'
+/** @deprecated Import `@dnb/eufemia/components/scroll-view/style` instead. This compatibility export will be removed in v13. */
+import '../../../components/scroll-view/style'
