+++ packages/dnb-eufemia/src/fragments/scroll-view/style/deps.scss
@@ -1,6 +1,1 @@
-/*
-* Used for snapshot testing
-*
-*/
-
-@use './dnb-scroll-view.scss';
+@use '../../../components/scroll-view/style/dnb-scroll-view.scss';
