+++ packages/dnb-eufemia/src/style/dnb-ui-components.scss
@@ -46,6 +46,7 @@
 @use '../components/popover/style/dnb-popover.scss';
 @use '../components/progress-indicator/style/dnb-progress-indicator.scss';
 @use '../components/radio/style/dnb-radio.scss';
+@use '../components/scroll-view/style/dnb-scroll-view.scss';
 @use '../components/section/style/dnb-section.scss';
 @use '../components/skeleton/style/dnb-skeleton.scss';
 @use '../components/skip-content/style/dnb-skip-content.scss';
